BUG: 337742 REVIEW: 120431. Fix and future-proof security in Dr Konqi.
In July 2014, the bugs.kde.org database switched to v4.4.5 of Bugzilla
and cookies were no longer accepted, causing bug 337742. This patch uses
the new security system, based on tokens, and implements the future use
of passwords only, whenever Bugzilla changes to the announced v4.5.x. It
also avoids a crash in Dr K on Apple OS X when KCookieJar is not found.

Forward port of review 120431 to plasma-workspace.

REVIEW: 120876