adjust the size of the window after the labels sizes were changed
Summary:
during backtrace loading the widgets get mutated a lot, changing their size
to very long, or very short, but certainly different than from what they
were originally when the window was sized.

as a result the window may now be too small to actually fit the content and
overlap some of the widgets. to resolve this problem we now force a size
adjustment when the backtracewidget changes. other pages, from what I
can tell are not affected.

the layout of this page is a fairly involved unfortuntely, so adjusting
on the fly is the best we can do I fear

CHANGELOG: The crash dialog no longer cuts off text after backtracing
BUG: 337319

Test Plan:
- crash
- try to report bug
- dialog text not cut off (LANGUAGE=de helps with triggering the problem)
- go to advanced backtrace view, still fine

- crash
- go to advanced view in the main window, still fine

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D19390