use even more resizing
Summary:
since the widget by default has no strings in the designer ui file any
state change injects stuff in qlabels and changes the effective size hint.
to deal with this once and for all the three state changes now always
call a new adjustment helper.
the adjustment helper calls adjustSize iff the hintted dimensions are
larger than the effective one. this way we don't shrink the window
when the backtrace is reloaded, but always grow it if needed for the
backtraced state.

CHANGELOG: The Developer Information tab now correctly resizes to fit the content
FIXED-IN: 5.16.1
BUG: 406748
BUGL 403408

Test Plan:
- no cut off widgets when going to dev info tab while bt is generated
- neither when the bt is loaded
- doesn't shrink when clicking reload

Reviewers: mart

Reviewed By: mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21642