use errorString instead of errorText
Summary:
the former is for user display, the latter is a sparse variant for
internal use. this results in nicer/useful error messages on errors

CCBUG: 405597
CHANGELOG: Error message when failing to save backtraces are more helpful.

Test Plan: - error is a readable string with description instead of only a file path now

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21644