copy backtrace file with overwrite flag
Summary:
QFileDialog was already run with confirmOverwrite, but we didn't actually
set the Overwrite flag on the copy call, so we got an error back when
the file already existed. Since the user already needed to confirm the
overwrite in the QFD we can always pass Overwrite.

BUG: 405597
CHANGELOG: Saving backtraces now overwrites files when necessary
FIXED-IN: 5.16.1

Test Plan:
- save backtrace
- save again
- dialog asks to confirm overwrite
- gets overwritten succesfully

- save again
- refuse to overwrite
- dialog stays open until cancel or new path is chosen
- no overwrite

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21645