show details examples in label itself instead of unexpected tooltip
Summary:
the previous UX wasn't good. the label contained a "link" highlight
but that link was only a trigger to a tooltip. IOW it behaved more like
a button than a link.

instead of using a tooltip fully expand the examples into an indented label
below the relevant checkbox.

BUG: 401697
FIXED-IN: 5.17.0

Reviewers: ngraham, #vdg

Reviewed By: ngraham, #vdg

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21745