do not set a hardcoded minimum size
Summary:
setting the minimum size manually like this is overriding the actual
calculated minimum size (as per child widgets) which ultimately means that
when a child's effective minimum is larger than the manually set one you
can resize the widget such that the child no longer fits in.
this is most notably observed with the page title widget. it doesn't word
wrap, so it has a specific minimum size: the amount of space it requires to
render the text in one line if its minimum width then exceeds the
600 pixels we had manually set, the widget would get cut off.

setting the minimum size should be entirely unnecessary. if all widgets
have a suitable sizing policy and sizehint the dialog will calculate a
suitable overall sizehint and minimum. getting rid of the manually set
minimum means the dialog can be shrunk exactly as far as the page content
allows it to shrink and no further.

I chased the the minimum size back to kde-runtime in git but it even seems
to predate that, so I have no clue what it was meant to achieve, but I am
almost certain that it was a hacky attempt at hiding bad size policies
or lack of size adjustments elsewhere in the dialog stack.

CHANGELOG: The bug report dialog can no longer be resized to cut off text
FIXED-IN: 5.16.3
BUG: 403408

Test Plan:
- set general font size to 16pt
- start drkonqi with LANGUAGE=pt_BR
- report bug
- observe all pages suffering from being cut off, up until the backtrace page which has manual adjustment logic as of a couple of commits ago (due to its actual sizing changing)

Reviewers: #plasma, apol

Reviewed By: apol

Subscribers: apol, cfeck, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21797