auto log in if username and password have been loaded from kwallet
Summary:
it's a bit moot to load the data from the wallet and then still have the
user manually click a log in button when we could just as well have tried
to log in using the stored credentials.

the new behavior is that the user goes through the bug reporting dialog,
and when showing the login page it will automatically try to login using
stored credentials. if the creds check out, the dialog automatically moves
on to the next page.

- page-validity is now in a helper canLogin function as it needs checking
  in two functions now (walletlogin and loginclick)
- the page now automatically goes into login state and attempts to log in
  using stored credentials

in the event that the credentials are wrong (e.g. password changed on
website) the page will enter error state saying the credentials are wrong.
when no credentials are in the wallet or the credentials are incomplete
(password remembering is off) no auto login happens.

BUG: 202495
FIXED-IN: 5.16.3
CHANGELOG: DrKonqi will now automatically log into bugs.kde.org when possible

Test Plan:
- have no credentials -> no auto login
- login manually and store credentials -> restart -> auto login happens
- change password on website -> restart -> auto login fails & error is shown

Reviewers: #plasma, ngraham

Reviewed By: ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D22190