change username labels to e-mail address instead
Summary:
users in bugzilla are always identified by their email address so it makes
no sense to label the input username.

calling it email address is also consistent with the bugzilla webui itself.

that said, we actually spell it inconstantly email or e-mail across
software. the most prominent cases in kdepim though spell it
"E-mail Address" for input labels. so, going with that

for ease of reading I'm leaving the code variables and the likes to say
username as this more obviously ties to the related password variables.

BUG: 375462
FIXED-IN: 5.17.0

Test Plan: login page now has an email address label

Reviewers: #plasma, ngraham, apol

Reviewed By: apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D22212