refactor basic OS info to be set via config and fixate it for testing
Summary:
the basic OS info is decided at compile time, on top of that
influences whether lsbrelease and osrelease are even run, and also
acts as seed value for some other members.
this resulted in tests failing on freebsd because it would detect freebsd
as build platform and then skip all the advanced lookup tech and having
unexpected values.

fetchOSBasicInformation is now a dangling static and used to determine
the default constructed config value of the operating system.
this is now also forced to read Linux for the tests.

I've also changed all the QLatin1Strings in the function to QStringLiteral.
as I recall QL1S is only faster when not turning the object into a
full QString, so these values shouldn't be QL1S.

Test Plan: tests pass on linux still (and hopefully freebsd CI too)

Reviewers: #plasma, apol

Reviewed By: apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D22328