update integration test
it seems the CI didn't run them in a whole while :/
fixate some calls as they dropped to the live bugzilla which since had
the bug states changed and then fails expectations. bring the fixtures
in line with expectations. this is slightly better in master already
as more stuff is fixated