always log raw exception data
Summary:
a UI representation of the error may lack all the data. specifically the
json blobs could be fairly long if perl hits a server-side exception
(e.g. bug in bugzilla itself). by logging it we have a way to get to
the raw data should it be necessary

Test Plan: builds and warns when appropriate

Reviewers: #plasma, apol

Reviewed By: apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D22398