pull platforms/distributions dynamically out of the bugzilla api
Summary:
this ensures that all known platforms are actually in the comobox and
reduces the amount of string duplication.

the new code gets the list of platforms from bugzilla and then either
attempts to select the one which was auto-detected by systeminformation
or recycle the one previously used.

setup is now triggered in the constructor. loading from the api is still
fairly cheap but even so takes longer than the hardcoded list, so
loading this on aboutToShow is fairly unattractive at this point.

the now useless members for combobox state tracking are gone entirely.

BUG: 383530
CHANGELOG: The list of available platforms is now always in sync with the list on bugzilla itself

Test Plan:
list gets loaded as expected
list errors as expected when offline

Reviewers: #plasma, apol

Reviewed By: apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D22492