Add option to discourage reporting bugs
This is for vendors and environments that don't want to allow their users to report bugs directly to us.
By setting InteractionAllowed to false in drkonqirc the following will change:

* For automatically restarted services, no hint is shown whatsoever
* On shutdown no hint is shown whatsoever
* For application crashes a notification is shown with a Restart button but no status icon nor "Report bug" button

The DrKonqi dialog itself is unchanged (for now?), which will show up in the rare case no notification service is running.
Also removed the icon name stuff that clearly never worked.

Differential Revision: https://phabricator.kde.org/D23261