fix default version selection to allow for renamed "unspecified" versions
Summary:
plasmashell for example has a "master" instead of "unspecified". if we
still tried to file a bug against "unspecified" bugzilla will throw rather
unhelpful errors. instead as a fallback to finding a "proper" version, make
sure to set the "default" to something sane. if the hardcoded default
doesn't map to an actual version as per the API we pick the lowest possible
id and file against that instead. seeing as you cannot remove versions this
should technically be the "unspecified" version no matter what it was
renamed to.

BUG: 411983

Test Plan: can file bugs against plasmashell even when kcrash passed an empty or unmapped version

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24214