further constrict line parsing of .so files
Summary:
As seen in https://bugs.kde.org/show_bug.cgi?id=417435 the previous fix
for .so detection actually falls flat on the face for most things
other than glib2. The original test here was flawed as libglib-2.0.so.0
has the complete suffix '0.so.0' which matched the original parsing
condition. most libraries however are libfoo.so.0 and for that the suffix
would be 'so.0' which did not match the original condition.

extend the condition to properly cover files without so-version, files with
awkward suffix (such as glib) and files with normal names and normal
versioning.
also extend test case accordingly

CCBUG: 418538

Test Plan: test passes. all lines from 417435 score less than `Good`

Reviewers: cfeck, ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28026