skip over _start and __libc_start_main
when an application has no debug symbols it will usually be missing the
functionName for main/kdemain and instead show the underlying libc helpers
in the trace. do not consider them in the rating, they add nothing of
value to the stack trace.

e.g. https://bugs.kde.org/show_bug.cgi?id=417435