remove pointless and arbitrary 4 line frame limit
Summary:
... and update a related test case

the original motivation here was garbage _start traces shouldn't be
considered valuable. of course we have way more advanced measurements for
whether a trace is valuable or not. notably the actual rating. so as
long as we have any frame for evaluation that frame's rating **is**
relevant.
drkonqi is smart enough to ignore lines that do not matter already.
e.g. everything above the crash handler is ignored, equally originator
functions main/start_thread/start/... are ignored. they do not impact
the rating at all, regardless of the amount of frames under review.
as such it is entirely possible to have fewer than 4 frames under review
and that those 4 frames are all we need (e.g. the crash is in main()
directly)

in fact, the very trace sample testing this was proving how the 4 line
limit made no sense. the frame about KCmdLineArgs is the only relevant
and only valuable frame in the sample and it is 100% useful making the
entire trace useful. it is simply a nullptr dereference in the main.

the frame limit is now 1 and the test sample has been replaced with
a real-life sample from https://bugs.kde.org/show_bug.cgi?id=193032

Test Plan: test passes again (was broken because _starts is now actively skipped)

Reviewers: apol, bcooksley

Reviewed By: apol

Subscribers: ahmadsamir, bcooksley, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28098