Don't leak DrKonqi dialog / fix crash on wayland
Summary:
When a QApplication is in it's destructor it calls window->destroy()
on all remaining QWindows. This deletes the platform window, whilst
keeping the real object alive. QtWayland doesn't like this.

Deleting the dialog fixes the issue. The deleteLater is fine because
QCoreApplication teardown is magic.

Test Plan: crashTest

Reviewers: #plasma, sitter

Reviewed By: sitter

Subscribers: anthonyfieroni, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28692