disable automatic cookie injection
the http slave reads the cookie metadata and defaults to auto, which would
try to find applicable cookies in the cookiejar and inject it into every
request. this is problematic because that cookie is not from drkonqi but
rather some other software (e.g. konqueror). as such the cookie is entirely
useless for us but gets in the way when bugzilla finds the cookie to be
unacceptable for whatever reason (outdated,invalid,unexpected).

force all apijobs to disable cookies entirely. we have no use for them.

BUG: 419646
FIXED-IN: 5.18.5