fix buttongroup state toggles
this was recently changed on the BugAwarenessPage and broke state
tracking as idPressed is emitted when pressing down, before the group
state actually switches. the yes/no group on that page was effectively
inverted as updateCheckBoxes was called before the property update and
thus when the user clicked Yes the internal state in the slot was still
No and vice versa.

the duplicates dialog similarly was wrong but it's much harder to
stumble over since the relevant UI elements are only visible when trying
to mark a bug duplicate to a closed bug or a "common" bug

use the idClicked signal instead, it's what the code originally used (as
the now deprecated buttonClicked)

BUG: 428346