Modify DBus service name to allow using QDBusServiceWatcher
Fix the DBus connection which allows passing a debuggee to one of the
KDevelop debuggers for future versions (>=5.7) of KDevelop. Instead of the
deprecated serviceOwnerChanged signal, a QDBusServiceWatcher is used,
which requires a different naming scheme. See KDevelop commit 69ef7ea8.

This change has no impact on the connection with current versions of KDevelop.