discard wayland client crashes following compositor crash
when the compositor goes down Qt will abort the clients causing kind of
erroneous crashes that the client can't do anything about. as such we
don't want to allow these reports from going through and instead
shortcut the report dialog to a final conclusion page where we inform
the user about the unfortunate circumstance surrounding this crash.

compositor crash detection is based on a simple check for the relevant
qfatalizing string resulting in a persistent state on the parser

BUG: 431561