include the drkonqi version in reports
when looking at odd ball reports that ended up on bugzilla I'm never
quite sure what version let the report get through. in the interest of
actually debugging defects with drkonqi we'll therefore want the drkonqi
version itself. this way it's easier to pin point a version that might
have (e.g.) problems detecting a trace as useless

also has the rather advantageous side effect that drkonqi is part of
plasma so its version will generally be the same as the rest of plasma
giving a sense of what versions are in play when plasma is involved