don't assert that there is 1 product
...there may be none!

we'll throw this as an exception to trigger error signaling in
bugzillalib.cpp. Alas, as it turns out this isn't actually wired up to
anything. so this effectively just turns the failed assert (which may
crash) into a runtimexception (which never crashes).

to reasonably deal with this the wizard will need some improvements as
well