wire up productInfoError to fall back to 'kde' product
when a product fails to resolve in bugzilla we'll want to fall back on
the generic 'kde' product so the user can still conveniently report the
bug and get it triaged.

this most notably deals with scenarios where a kaboutdata product/binary
does not have its own bugzilla product and also has no mapping data in
our file (e.g. kdeinit5 is part of the frameworks-kinit product, without
mapping data we'll not able to file it there and the submission fails)

when product mapping now fails it will signal an error, the error is
connected to the fallback logic, which simply switches the product to
'kde' and retries the resolution

there's also a hint about why this bug was filed against 'kde' being
added to the report summary so people don't get confused and also to
perhaps prompt them to fix the mappings file


(cherry picked from commit 6360b56bcc8a68054723f3e555f5eb1798964ebd)