refactor deleted file detection
instead of using procfs' /map_files use /maps. this will allow us to
also process data after the process had gone away (e.g. when processing
data from coredumpd) and makes no difference for the running process.

the parser is now in its own file and has a unit test

there's also new tech in place that additionally adds stat support. this
is also to aid in future development for coredumpd support. if a library
has changed in the time between the core dumping and drkonqi starting,
then we need a way to figure out if the file has changed even when we
can't rely on the ' (deleted)' marker in the /maps data.
the easiest way to achieve this is to compare the inodes. while that
isn't always accurate it's probably accurate enough for our purposes.