disable dbus interface and explain what it does
this wasn't really obvious from the get go, but from the API I'd guess
it's so a running kdevelop can register itself as external debugger on a
running drkonqi. we limit this to kcrash (for now anyway) there's not
really a reason kdevelop would have to route coredumpd dumps through
drkonqi and it'd needlessly complicate things, in particular since we
need coredumpbackend to have the prepare API that kdevelop would have to
implement to use the backend over dbus, so the interface also needs to
grow api, and then the interface also needed to grow API to communicate
backend support.