prepare the backend before tracing
debuggermanager now plays negotiator between backend and generator so
that the backend can prepare itself for tracing.
specifically this will allow us to interface with coredumpd where we
might need to extract the dump for tracing just before starting the
trace. this will be possible with the new preparation dance between
generator, manager and backend.

instead of outright starting the generator now enters loading state and
emits `preparing()` to get the backend to prepare. once the backend is
prepared it in turn emits a signal to trigger further startup in the
generator.